# docs/openapi/overlay-web.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Overlay Web API",
    "version": "1.0.0",
    "description": "Generated from the overlay-web route boundary registry. This first-pass reference documents stable route inventory and request shapes."
  },
  "servers": [
    {
      "url": "https://getoverlay.io",
      "description": "Hosted Overlay web app"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development web app"
    }
  ],
...
  "paths": {
...
    "/api/v1/generate-image": {
      "post": {
        "operationId": "postGenerateImage",
        "summary": "Generate an image",
        "tags": [
          "Media"
        ],
        "security": [
          {
            "browserSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response. Stable response schemas are documented in workflow guides when available.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Response shape is route-specific and may be backed by @overlay/app-core contracts."
                }
              }
            }
          },
          "400": {
            "description": "Invalid request.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid authentication.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "accessToken": {
                    "type": "string",
                    "minLength": 1
                  },
                  "userId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "prompt": {
                    "type": "string",
                    "minLength": 1
                  },
                  "modelId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "aspectRatio": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 20
                  },
                  "conversationId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "turnId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "imageUrl": {
                    "type": "string",
                    "minLength": 1
                  },
                  "temporaryChat": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "browserSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "overlay_session",
        "description": "Signed httpOnly browser session cookie."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "WorkOS bearer access token for native or service clients."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": true
      }
    }
  }
}